Completed Medium 1493. Longest Subarray of 1's After Deleting One Element
This method is using 3 index tracking ints and a max variable. I am keeping track of the start and end of the subarray. I am also keeping track of the item to delete if it is within the subarray. This extra variable lets me use logic to adjust the new subarray, when a 0 is hit instead of reading until a 0 is hit. I initialize the index tracking variables to the opposite side of the array to be able to know if they are set (ie. start of the end and end before the start).