Completed Medium 274. H-Index
This is the approach that first came to mind. put all of the numbers in a heap so that they are sorted. once they are all on the heap you can see if the top number is smaller than the count and de-queue it until its not. Then, having to deal with H values that are not in the array, taking the max of the last popped off and the length of the heap.