Completed Hard 135. Candy
This was an intuitive algo. The premis is, we keep a running counter if its increasing. Once we reach a peak, we mark it and save what the peak would be for this node. Once we reach the bottom (or two equal side by side) of the priorities, we get how many numbers are involved (n). Add Min(peak,n) and then add the exclusive natural sum of n.

natural sum of n inclusive is (n(n+1))/2
natural sum of n exclusive is (n(n-1))/2