Completed Hard 42. Trapping Rain Water
Completed by a 2 pointer method. We find the first left max, then the right max. We get add the water level for each array element while it is less than both of those. If we find a new left peak, we move the rigth pointer and continue. The opposite if we are moving the right side and find a right peak.