Completed Medium 12. Integer to Roman
This is a simple solution. Subtract as much as possible from the int value without being negitive and append the corrisponding string to the string builder. loop through the values from greatest to smallest.

This looks like it would be O(n^2), but the for is looping O(1). The while has linear growth, since 2000000 will not take the same iterations as 2.

 Over all time complexity is O(n) and space complexity of O(1);