Completed Medium 1647. Minimum Deletions to Make Character Frequencies Unique
Used an intuative method. Time Complexity O(n+m) Space Complexity O(1).

n is the string length
m is the number of times that are needed to decrament each frequency count in the while loop. This can become a non-trivial amount in cases some cases. For example, when each char has the same frequency.

1. get the frequencies
2. Try and insert all of them into a hash set.
3. if the hasthe key already, decrament the frequency, incrament the altercation count and try step 2 again.