ListNode bug and ToList feature
Added a ToList extension method on ListNodes. This will make unit testing easier by allowing a .ToList().ToArray(). Test performance is not a big issue. If it becomes one, we can scan the list and create a perfectly sized array/list.

Int[] -> ToListNode had a bug where there would be empty Listnode cells in the output.